backend: expand file hierarchy, add config file
Until now, the FastAPI settings were hard-coded in the main.py file.

With this commit, the settings are set in the config.py file, based on
the API_CONFIG environment variable. If not set, it will fallback to
default which is "development".

This helps to switch from one environment (e.g. development) to
another (e.g. testing, production) without the need to worry about
changing ip, database name, etc.